Network Propagation of PBI marker changes WT
This code projects the PBI markers changing upon HU-induced DNA replication stress in wild type cells identified with analysis pipeline 2 onto a protein-protein interaction network and extracts regions from the network that are closely related to the hits. Clustering is applied to identify protein complexes that are likely to undergo changes in assembly state under this condition.